+++ py/desitarget/mtl.py
@@ -252,15 +252,27 @@ def get_mtl_tile_file_name(secondary=False):
     return fn
 
 
-def get_ztile_file_name():
+def get_ztile_file_name(survey='main'):
     """Convenience function to grab the name of the ZTILE file.
+
+    survey : :class:`str`, optional, defaults to "main"
+        To look up the correct ZTILE filename. Options are ``'main'`` and
+        ``'svX``' (where X is 1, 2, 3 etc.) for the main survey and
+        different iterations of SV, respectively.
 
     Returns
     -------
     :class:`str`
         The name of the ZTILE file.
     """
-    fn = "tiles.csv"
+    if survey[:2] == 'sv':
+        fn = "tiles.csv"
+    elif survey == 'main':
+        fn = "tiles-specstatus.ecsv"
+    else:
+        msg = "Allowed 'survey' inputs are sv(X) or main, not {}".format(survey)
+        log.critical(msg)
+        raise ValueError(msg)
 
     return fn
 
@@ -1011,7 +1023,7 @@ def tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey):
     Parameters
     ----------
     zcatdir : :class:`str`
-        Full path to the "daily" directory that hosts redshift catalogs.
+        Full path to the directory that hosts redshift catalogs.
     mtltilefn : :class:`str`
         Full path to the file of tiles that have been processed by MTL.
     obscon : :class:`str`
@@ -1028,9 +1040,27 @@ def tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey):
     :class:`~numpy.array`
         An array of tiles that have not yet been processed and written to
         the mtl tile file.
+
+    Notes
+    -----
+    - If `survey` is `'main'` the code assumes the file with zdone for
+      spectro tiles is `mtltilefn`/../../ops/tiles-specstatus.ecsv
+      (i.e. it is in the ops directory parallel to the mtl directory).
+      If `survey` is `'svX'` the code assumes it is `zcatdir`/tiles.csv.
     """
     # ADM read in the ZTILE file.
-    ztilefn = os.path.join(zcatdir, get_ztile_file_name())
+    ztilefn = get_ztile_file_name(survey=survey)
+    # ADM directory structure is different for SV and the Main Survey.
+    if survey[:2] == 'sv':
+        ztilefn = os.path.join(zcatdir, ztilefn)
+    elif survey == 'main':
+        opsdir = os.path.dirname(mtltilefn).replace("mtl", "ops")
+        ztilefn = os.path.join(opsdir, ztilefn)
+    else:
+        msg = "Allowed 'survey' inputs are sv(X) or main, not {}".format(survey)
+        log.critical(msg)
+        raise ValueError(msg)
+
     tilelookup = Table.read(ztilefn)
 
     # ADM the ZDONE column is a string, convert to a Boolean.
@@ -1211,6 +1241,11 @@ def make_zcat_rr_backstop(zcatdir, tiles, obscon, survey):
             allzs.append(zz)
             # ADM read in all of the exposures in the fibermap.
             fm = fitsio.read(zbestfn, "FIBERMAP")
+            # ADM in the transition between SV3 and the Main Survey, the
+            # ADM fibermap data model changed. New columns may need to be
+            # ADM removed to concatenate old- and new-style fibermaps.
+            if "PLATE_RA" in fm.dtype.names:
+                fm = rfn.drop_fields(fm, ["PLATE_RA", "PLATE_DEC"])
             # ADM recover the information for unique targets based on the
             # ADM first entry for each TARGETID.
             _, ii = np.unique(fm['TARGETID'], return_index=True)
@@ -1284,9 +1319,9 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
         Options are ``'main'`` and ``'svX``' (where X is 1, 2, 3 etc.)
         for the main survey and different iterations of SV, respectively.
     zcatdir : :class:`str`, optional, defaults to ``None``
-        Full path to the "daily" directory that hosts redshift catalogs.
-        If this is ``None``, look up the redshift catalog directory from
-        the $ZCAT_DIR environment variable.
+        Full path to the directory that hosts redshift catalogs. If this
+        is ``None``, look up the redshift catalog directory from the
+        $ZCAT_DIR environment variable.
     mtldir : :class:`str`, optional, defaults to ``None``
         Full path to the directory that hosts the MTL ledgers and the MTL
         tile file. If ``None``, then look up the MTL directory from the
@@ -1306,8 +1341,8 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     :class:`str`
         The name of the MTL tile file that was updated.
     :class:`str`
-        The name of the ZTILE file that was used to link TILEIDs to
-        observing conditions and to determine if tiles were "done".
+        Name of ZTILE file used to link TILEIDs to observing conditions
+        to determine if tiles were "done" (that they had zdone=True).
     :class:`~numpy.array`
         Information for the tiles that were processed.
 
@@ -1315,6 +1350,10 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     -----
     - Assumes all of the relevant ledgers have already been made by,
       e.g., :func:`~desitarget.mtl.make_ledger()`.
+    - If `survey` is `'main'` the code assumes the file with the zdone
+      status for spectro tiles is `mtldir`/../ops/tiles-specstatus.ecsv
+      (i.e. it is in the ops directory parallel to the mtl directory).
+      If `survey` is `'svX'` the code assumes it is `zcatdir`/tiles.csv.
     """
     # ADM first grab all of the relevant files.
     # ADM grab the MTL directory (in case we're relying on $MTL_DIR).
@@ -1336,11 +1375,18 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
                                      survey=survey, obscon=obscon, ender=form)
     # ADM grab the zcat directory (in case we're relying on $ZCAT_DIR).
     zcatdir = get_zcat_dir(zcatdir)
-    # ADM And contruct the associated ZTILE filename.
-    ztilefn = os.path.join(zcatdir, get_ztile_file_name())
 
     # ADM grab an array of tiles that are yet to be processed.
     tiles = tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey)
+
+    # ADM contruct the ZTILE filename, for logging purposes.
+    ztilefn = get_ztile_file_name(survey=survey)
+    # ADM directory structure is different for SV and the Main Survey.
+    if survey[:2] == 'sv':
+        ztilefn = os.path.join(zcatdir, ztilefn)
+    elif survey == 'main':
+        opsdir = os.path.dirname(mtltilefn).replace("mtl", "ops")
+        ztilefn = os.path.join(opsdir, ztilefn)
 
     # ADM stop if there are no tiles to process.
     if len(tiles) == 0: